+++ src/main/java/edu/gsu/bbb/willdo/Controllers/GroupController.java
@@ -1,6 +1,7 @@
-package edu.gsu.bbb.willdo;
+package edu.gsu.bbb.willdo.Controllers;
 
-import org.bson.types.ObjectId;
+import edu.gsu.bbb.willdo.models.Group;
+import edu.gsu.bbb.willdo.Repositories.GroupRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
