+++ src/main/java/edu/gsu/bbb/willdo/Controllers/TaskController.java
@@ -1,6 +1,9 @@
-package edu.gsu.bbb.willdo;
+package edu.gsu.bbb.willdo.Controllers;
 
-import java.util.*;	
+import java.util.*;
+
+import edu.gsu.bbb.willdo.models.Task;
+import edu.gsu.bbb.willdo.Repositories.TaskRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.http.HttpStatus;
