+++ src/main/java/edu/gsu/bbb/willdo/Repositories/GroupRepository.java
@@ -1,5 +1,6 @@
-package edu.gsu.bbb.willdo;
+package edu.gsu.bbb.willdo.Repositories;
 
+import edu.gsu.bbb.willdo.models.Group;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface GroupRepository extends MongoRepository<Group,String> {
