+++ src/main/java/edu/gsu/bbb/willdo/Repositories/TaskRepository.java
@@ -1,5 +1,6 @@
-package edu.gsu.bbb.willdo;
+package edu.gsu.bbb.willdo.Repositories;
 
+import edu.gsu.bbb.willdo.models.Task;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.util.List;
