+++ src/main/java/edu/gsu/bbb/willdo/models/Task.java
@@ -1,4 +1,4 @@
-package edu.gsu.bbb.willdo;
+package edu.gsu.bbb.willdo.models;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.index.Indexed;
